requirements.txt
pytz~=2022.1
requests==2.28.1
beautifulsoup4~=4.11.1
python-dotenv~=0.20.0
Flask~=2.1.2
Jinja2~=3.1.2
firebase-admin~=6.1.0